Kinder Garden/requirements.txt
fastapi==0.115.0
sqlalchemy==2.0.35
pydantic==2.9.2
uvicorn==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.6
redis==5.1.0
flower==2.0.1
